[Decomp][Player] Cohesion split: extract load system to PlayerLoad.cpp
